pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nexuscargo-platform"
version = "0.1.0"
description = "NexusCargo AI Platform"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn>=0.35.0",
  "sqlalchemy>=2.0.36",
  "alembic>=1.14.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.7.0",
  "python-jose[cryptography]>=3.3.0",
  "httpx>=0.28.1",
  "python-multipart>=0.0.12",
  "google-cloud-pubsub>=2.29.0",
  "google-cloud-storage>=2.18.2",
  "google-cloud-secret-manager>=2.21.1",
  "redis>=5.2.1",
  "tenacity>=9.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4",
  "pytest-cov>=6.0.0",
  "mypy>=1.14.1",
  "ruff>=0.9.6"
]

[tool.setuptools]
packages = []

[tool.ruff]
line-length = 100
target-version = "py39"
exclude = ["apps/web", "infra", ".venv", "node_modules"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501", "B008", "UP045"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --cov=services.ingestion.service --cov=services.review.service --cov=services.validation.service --cov=services.classification.service --cov=services.extraction.service --cov=libs.common.idempotency --cov=libs.common.audit --cov=libs.auth.security --cov=modules.awb.service --cov=modules.fiar.service --cov-report=term-missing --cov-fail-under=80"

[tool.mypy]
python_version = "3.9"
strict = true
exclude = "(apps/web|infra|alembic)"
ignore_missing_imports = true
warn_unused_configs = true
